requirements.txt
# Python Dependencies for SQL Agent Demo

# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Azure OpenAI
openai==1.12.0

# Database
pyodbc>=5.2.0

# Azure Authentication
azure-identity==1.15.0

# Environment and Configuration
python-dotenv==1.0.0

# HTTP Client
requests==2.31.0
aiohttp==3.9.0

# Note: POML (Prompt Optimization Markup Language) is integrated via inline prompts
# Reference: https://microsoft.github.io/poml/latest/
